refactor: merged encryption package with authentication
