+++ pkg/authentication/cookies_test.go
@@ -15,7 +15,6 @@ import (
 
 //go:generate mockgen -build_flags=--mod=mod -package authentication -destination ./mock_logger.go -source=../../internal/logging/interfaces.go
 //go:generate mockgen -build_flags=--mod=mod -package authentication -destination ./mock_interfaces.go -source=./interfaces.go
-//go:generate mockgen -build_flags=--mod=mod -package authentication -destination ./mock_encrypt.go -source=../../internal/encryption/interfaces.go
 //go:generate mockgen -build_flags=--mod=mod -package authentication -destination ./mock_tracing.go go.opentelemetry.io/otel/trace Tracer
 //go:generate mockgen -build_flags=--mod=mod -package authentication -destination ./mock_monitor.go -source=../../internal/monitoring/interfaces.go
